playing aroung with using a makefile
Why? Because it's more standard, and I'd like to be able to run tests
without using a custom script and half-hacking vows. Thus, `npm test`
now works.

This neccessitated a change in the way exports works, so there is now a
`head` and `tail` that are concatenated. They are not valid JS on their
own and will throw errors is loaded manually.

Loading each file manually in a browser (such as `test/test.html`) still
works just fine, provided that `var exports = vash = {}` is defined
globally. Obviously this is only for development.